requirements.txt
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
appnope==0.1.4
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
daff==1.3.46
dbt-adapters==1.14.6
dbt-common==1.22.0
dbt-core==1.9.4
dbt-duckdb==1.9.2
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.7.4
dbt-snowflake==1.9.2
debugpy==1.8.14
decorator==5.2.1
deepdiff==7.0.1
duckdb==1.2.2
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.8
leather==0.4.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.1
matplotlib-inline==0.1.7
more-itertools==10.6.0
mplcursors==0.6
msgpack==1.1.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.11.5
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.14.1
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.45.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
zipp==3.21.0
